Separating project into a server and client module.
