+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/
+lib/
 logs/
 docs/
 .DS_Store
