+++ server/routes/profile.js
@@ -1,14 +1,14 @@
 const express = require('express');
 const router = express.Router();
 
-import parse from '../parser/stats';
+import parse from '../../src/parser/profile';
 import cache from '../cache';
 import utils from '../utils';
 
 /**
- * @api {get} /stats/:platform/:region/:tag Get profile of player.
- * @apiName GetStats
- * @apiGroup Stats
+ * @api {get} /profile/:platform/:region/:tag Get profile of player.
+ * @apiName GetProfile
+ * @apiGroup Profile
  *
  * @apiParam {String} platform Name of user. pc/xbl/psn
  * @apiParam {String} region Region of player. us/eu/kr/cn/global
@@ -17,15 +17,33 @@ import utils from '../utils';
  * @apiSuccess {Object} data Profile data.
  *
  * @apiExample {curl} Example usage:
- *  curl -i http://ow-api.herokuapp.com/stats/pc/us/user-12345
+ *  curl -i http://ow-api.herokuapp.com/profile/pc/us/user-12345
  *
  * @apiSuccessExample {json} Success-Response:
     HTTP/1.1 200 OK
     {
-      username: "user"
-      stats: {
-        top_heroes: {...}
-        combat: {...}
+     data: {
+        username: "user",
+        games: {
+          quickplay: {
+            wins: "252"
+          },
+          competitive: {
+            wins: "9",
+            lost: 18,
+            played: "27"
+          }
+        },
+        playtime: {
+          quickplay: "63 hours",
+          competitive: "5 hours"
+        },
+        competitive: {
+          rank: "2083",
+          rank_img: "https://blzgdapipro-a.akamaihd.net/game/rank-icons/rank-10.png"
+        },
+        levelFrame: "https://blzgdapipro-a.akamaihd.net/game/playerlevelrewards/0x025000000000091F_Border.png",
+        star: ""
       }
     }
  */
@@ -36,10 +54,10 @@ router.get('/:platform/:region/:tag', (req, res) => {
   const tag = req.params.tag;
   const include = req.query.include && req.query.include.split(',') || null;
 
-  const cacheKey = `stats_${platform}_${region}_${tag}`;
+  const cacheKey = `profile_${platform}_${region}_${tag}`;
   const timeout = 60 * 5; // 5 minutes.
 
-  cache.getOrSet(cacheKey, timeout, getStats, function(data) {
+  cache.getOrSet(cacheKey, timeout, getProfile, function(data) {
     if (data.statusCode) {
       res.status(data.response.statusCode).send(data.response.statusMessage);
     } else {
@@ -48,7 +66,7 @@ router.get('/:platform/:region/:tag', (req, res) => {
     }
   });
 
-  function getStats(callback) {
+  function getProfile(callback) {
     parse(platform, region, tag, (data) => {
       callback(data);
     });
