+++ src/parser/stats.js
